+++ client/src/components/AddNewArtwork/AddNewArtwork.css
@@ -1,0 +1,18 @@
+.sct_add_form {
+  border-radius: 20px;
+  box-shadow: -2px -2px 2.2px rgb(255 255 255 / 7%);
+  min-height: 50%;
+}
+
+.artworkForm {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+}
+/* 
+@media screen and (min-width: 768px) and (max-width: 1023px) {
+}
+@media screen and (min-width: 1024px) {
+} */
