+++ client/src/components/Geocoding/Geocoding.css
@@ -1,0 +1,95 @@
+input[type="search"] {
+  font-family: var(--text-font);
+  font-weight: 700;
+  padding: 0.8em 1em;
+}
+
+.search-form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 5em 0;
+}
+
+.search-btn {
+  height: 2em;
+  width: 8em;
+  align-self: center;
+  font-family: var(--text-font);
+  font-size: 1em;
+  font-weight: 700;
+  background-color: var(--jet-color);
+  color: var(--mustard-color);
+  border: solid 1px var(--mustard-color);
+  border-radius: 12px;
+}
+.search-btn {
+  cursor: pointer;
+}
+.search-btn:hover {
+  scale: calc(1.02);
+}
+
+.searchBar {
+  display: flex;
+  flex-direction: column;
+}
+.searchGeo {
+  display: flex;
+  flex-direction: column;
+}
+.citySearch {
+  width: 20em;
+}
+
+@media screen and (min-width: 768px) and (max-width: 1023px) {
+  form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+  }
+  .searchBar {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+  }
+  .search-btn {
+    background-color: var(--mustard-color);
+    color: var(--jet-color);
+    align-self: baseline;
+    margin: 0 2em;
+  }
+  .searchGeo {
+    display: flex;
+    flex-direction: column;
+  }
+  .citySearch {
+    width: 30em;
+  }
+}
+
+@media screen and (min-width: 1024px) {
+  .citySearch {
+    width: 40em;
+  }
+  .searchBar {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+  }
+  .search-btn {
+    background-color: var(--mustard-color);
+    color: var(--jet-color);
+    align-self: baseline;
+    margin: 0 2em;
+  }
+  .searchGeo {
+    display: flex;
+    flex-direction: column;
+  }
+}
