+++ client/src/components/Geolocalisation/Geolocalisation.css
@@ -1,0 +1,14 @@
+.geo-btn {
+  border: transparent;
+  background-color: transparent;
+  color: var(--mustard-color);
+  font-family: var(--text-font);
+  font-size: 1em;
+  margin: .5em 0;
+  font-weight: 500;
+  cursor: pointer;
+}
+
+.geo-btn:hover {
+  scale: calc(1.02);
+}
