+++ client/src/components/SearchBar/SearchBar.css
@@ -1,15 +1,3 @@
-form {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-}
-input[type="search"] {
-  font-family: var(--text-font);
-  font-weight: 700;
-  padding: 0.8em 1em;
-}
 .search-sct {
   display: flex;
   flex-direction: column;
@@ -17,101 +5,9 @@ input[type="search"] {
   align-items: center;
   /* padding: 5em 0; */
 }
-.search-form {
-  display: flex;
-  justify-content: center;
-  /* align-items: flex-end; */
-  padding: 5em 0;
-}
-.geo-btn {
-  border: transparent;
-  background-color: transparent;
-  color: var(--mustard-color);
-  font-family: var(--text-font);
-  font-size: 1em;
-  margin: .5em 0;
-  font-weight: 500;
-}
-.search-btn {
-  height: 2em;
-  width: 8em;
-  align-self: center;
-  font-family: var(--text-font);
-  font-size: 1em;
-  font-weight: 700;
-  background-color: var(--jet-color);
-  color: var(--mustard-color);
-  border: solid 1px var(--mustard-color);
-  border-radius: 12px;
-}
-.search-btn,
-.geo-btn {
-  cursor: pointer;
-}
-.search-btn:hover,
-.geo-btn:hover {
-  scale: calc(1.02);
-}
 
-.searchBar {
-  display: flex;
-  flex-direction: column;
-}
-.searchGeo {
-  display: flex;
-  flex-direction: column;
-}
-.citySearch {
-  width: 20em;
-}
-
-@media screen and (min-width: 768px) and (max-width: 1023px) {
-  form {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-  }
-  .searchBar {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-  }
-  .search-btn {
-    background-color: var(--mustard-color);
-    color: var(--jet-color);
-    align-self: baseline;
-    margin: 0 2em;
-  }
-  .searchGeo {
-    display: flex;
-    flex-direction: column;
-  }
-  .citySearch {
-    width: 30em;
-  }
+/* @media screen and (min-width: 768px) and (max-width: 1023px) {
 }
 
 @media screen and (min-width: 1024px) {
-  .citySearch {
-    width: 40em;
-  }
-  .searchBar {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-  }
-  .search-btn {
-    background-color: var(--mustard-color);
-    color: var(--jet-color);
-    align-self: baseline;
-    margin: 0 2em;
-  }
-  .searchGeo {
-    display: flex;
-    flex-direction: column;
-  }
-}
+} */
