+++ client/src/pages/NewArtwork/NewArtwork.css
@@ -1,0 +1,25 @@
+.add_artwork {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20% 10% 3% 10%;
+}
+
+.presentation_add {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1em;
+}
+
+.thankyou {
+  color: var(--mustard-color);
+}
+
+/* @media screen and (min-width: 768px) and (max-width: 1023px) {
+}
+
+@media screen and (min-width: 1024px) {
+} */
