+++ server/src/router.ts
@@ -6,11 +6,12 @@ const router = express.Router();
 // Define Your API Routes Here
 /* ************************************************************************* */
 
-import artworkActions from "./modules/item/artwork/artworkActions";
+import artworkActions from "./modules/artwork/artworkActions";
 import nominatimActions from "./nominatimActions";
 
 router.get("/api/artworks", artworkActions.browse);
 router.get("/api/artworks/:id", artworkActions.read);
+router.put("/api/artwork", artworkActions.add);
 router.get("/api/geolocalisation", nominatimActions.geocode);
 
 /* ************************************************************************* */
